+++ examples/functions/combinations/function.tf
@@ -1,0 +1,7 @@
+# Return the number of combinations of from 3 select 2:
+provider::stdlib::combinations(3, 2)
+# result => 3
+
+# Return the number of combinations of from 4 select 5:
+provider::stdlib::combinations(4, 5)
+# result => 0